manage script
