--- docker-compose.yml
@@ -1,22 +1,0 @@
-data:
-    image: busybox
-    volumes:
-        - ./dbdata:/dbdata/data
-
-db:
-    image: postgres
-    volumes_from:
-        - data
-    environment:
-        POSTGRES_PASSWORD: docker
-        PGDATA: /dbdata/data
-
-web:
-    build: .
-    command: python run.py -p 3031
-    volumes:
-        - .:/osrc
-    ports:
-        - "3031:3031"
-    links:
-        - db
