+++ environment.yml
@@ -1,20 +1,26 @@
 name: osrc
 dependencies:
+- certifi=2015.9.6.2=py35_0
 - flask=0.10.1=py35_1
 - itsdangerous=0.24=py35_0
 - jinja2=2.8=py35_0
 - markupsafe=0.23=py35_0
-- openssl=1.0.1k=1
+- openssl=1.0.2d=0
 - pip=7.1.2=py35_0
 - psycopg2=2.6.1=py35_0
-- python=3.5.0=0
+- python=3.5.0=1
 - readline=6.2=2
+- requests=2.8.1=py35_0
 - setuptools=18.4=py35_0
 - sqlalchemy=1.0.9=py35_0
 - sqlite=3.8.4.1=1
 - tk=8.5.18=0
+- tornado=4.2.1=py35_0
 - werkzeug=0.10.4=py35_0
 - wheel=0.26.0=py35_1
 - xz=5.0.5=0
 - zlib=1.2.8=0
+- pip:
+  - flask-script==2.0.5
+  - flask-sqlalchemy==2.1
