--- requirements.txt
@@ -1,1 +1,0 @@
-Flask-SQLAlchemy==2.1
