+++ run.py
@@ -1,38 +1,33 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-import os
 import argparse
 
 from tornado.ioloop import IOLoop
 from tornado.wsgi import WSGIContainer
 from tornado.httpserver import HTTPServer
 
-try:
-    import osrc  # NOQA
-except ImportError:
-    import sys
-    sys.path.insert(
-        0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-finally:
-    from osrc import create_app
+from werkzeug.serving import run_simple
+
+from osrc import create_app
 
 if __name__ == "__main__":
-    dirname = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-    default_settings = os.path.join(dirname, "local.py")
-
     parser = argparse.ArgumentParser()
     parser.add_argument("-p", "--port", default=3031, type=int,
                         help="the port to expose")
+    parser.add_argument("-d", "--debug", action="store_true",
+                        help="debugging interface")
     parser.add_argument("-f", "--filename",
                         default=None,
                         help="a Python file with the app settings")
     args = parser.parse_args()
+    print("port: {0}".format(args.port))
 
-    # Build the Flask app.
     app = create_app(args.filename)
-
-    # Fire up the tornado server.
-    http_server = HTTPServer(WSGIContainer(app))
-    http_server.listen(args.port)
-    IOLoop.instance().start()
+    if args.debug:
+        run_simple("0.0.0.0", args.port, app, use_reloader=True,
+                   use_debugger=True)
+    else:
+        http_server = HTTPServer(WSGIContainer(app))
+        http_server.listen(args.port)
+        IOLoop.instance().start()
